requirements.txt
torch
torchvision
tqdm
Pillow
opencv-python
